python/python_web/django_rest_framework/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==2.2
djangorestframework==3.12.2
mysqlclient==2.0.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.4.1
wincertstore==0.2
